Add .copr/Makefile to work around rpmautospec+COPR limitations
rpmautospec doesn't really work properly outside of dist-git. So instead
this Makefile replaces %{autorelease} with the current UNIX timestamp.
Quite hacky and weird-looking, but it achieves the desired result of
ensuring that successive COPR builds of a package have a larger release
number than the ones that came before. Hopefully.

Signed-off-by: Sam Day <me@samcday.com>